Completed exercise

+++ chapter1/exercise-conversion.go
@@ -5,7 +5,7 @@
  * the value you want to convert in parentheses.
  * ------------------------------------------------------------------------
  */
- 
+
 package main
 
 import "fmt"
@@ -15,13 +15,13 @@ func main() {
 	fmt.Println("Price is", price, "dollars.")
 
 	var taxRate float64 = 0.08
-	var tax float64 = price * taxRate
+	var tax float64 = float64(price) * taxRate
 	fmt.Println("Tax is", tax, "dollars.")
 
-	var total float64 = price + taxRate
+	var total float64 = float64(price) + taxRate
 	fmt.Println("Total cost is", total, "dollars.")
 
 	var availableFunds int = 120
 	fmt.Println(availableFunds, "dollars available.")
-	fmt.Println("Within budget?", total <= availableFunds)
+	fmt.Println("Within budget?", total <= float64(availableFunds))
 }
